// src/app/core/api/generated/models/cart-item-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CartItemDto {
  id: string;
  image_url: string;
  product_id: string;
  product_name: string;
  product_slug: string;
  quantity: (number | string);
  subtotal: (number | string);
  unit_price: (number | string);
}